Split out API from RPC.
