+++ core/rpc/Cargo.toml
@@ -6,28 +6,20 @@ edition = "2018"
 
 [dependencies]
 api = { package = "substrate-rpc-api", path = "./api" }
-derive_more = "0.14.0"
-futures = "0.1"
-futures03 = { package = "futures-preview", version = "0.3.0-alpha.17", features = ["compat"] }
-jsonrpc-core = "13.0.0"
-jsonrpc-core-client = "13.0.0"
-jsonrpc-pubsub = "13.0.0"
-jsonrpc-derive = "13.0.0"
-log = "0.4"
-parking_lot = "0.9.0"
-codec = { package = "parity-scale-codec", version = "1.0.0" }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
 client = { package = "substrate-client", path = "../client" }
-substrate-executor = { path = "../executor" }
-network = { package = "substrate-network", path = "../network" }
+codec = { package = "parity-scale-codec", version = "1.0.0" }
+futures03 = { package = "futures-preview", version = "0.3.0-alpha.17", features = ["compat"] }
+log = "0.4"
 primitives = { package = "substrate-primitives", path = "../primitives" }
-session = { package = "substrate-session", path = "../session" }
-state_machine = { package = "substrate-state-machine", path = "../state-machine" }
-transaction_pool = { package = "substrate-transaction-pool", path = "../transaction-pool" }
-sr-primitives = { path = "../sr-primitives" }
+rpc = { package = "jsonrpc-core", version = "13.0.0" }
 runtime_version = { package = "sr-version", path = "../sr-version" }
+serde_json = "1.0"
+session = { package = "substrate-session", path = "../session" }
+sr-primitives = { path = "../sr-primitives" }
+state_machine = { package = "substrate-state-machine", path = "../state-machine" }
+substrate-executor = { path = "../executor" }
 substrate-keystore = { path = "../keystore" }
+transaction_pool = { package = "substrate-transaction-pool", path = "../transaction-pool" }
 
 [dev-dependencies]
 assert_matches = "1.1"
