+++ core/rpc/api/Cargo.toml
@@ -7,17 +7,15 @@ edition = "2018"
 [dependencies]
 codec = { package = "parity-scale-codec", version = "1.0.0" }
 derive_more = "0.14.0"
-futures = "0.1"
 futures03 = { package = "futures-preview", version = "0.3.0-alpha.17", features = ["compat"] }
-jsonrpc-core = "13.0.0"
 jsonrpc-core-client = "13.0.0"
 jsonrpc-derive = "13.0.0"
 jsonrpc-pubsub = "13.0.0"
 log = "0.4"
 parking_lot = "0.9.0"
 primitives = { package = "substrate-primitives", path = "../../primitives" }
+rpc = { package = "jsonrpc-core", version = "13.0.0" }
 runtime_version = { package = "sr-version", path = "../../sr-version" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-sr-primitives = { path = "../../sr-primitives" }
 txpool = { package = "substrate-transaction-graph", path = "../../transaction-pool/graph" }
