+++ core/rpc/api/src/errors.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Substrate.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::rpc;
+use rpc;
 use log::warn;
 
 pub fn internal<E: ::std::fmt::Debug>(e: E) -> rpc::Error {
