+++ core/rpc/api/src/state/error.rs
@@ -14,35 +14,53 @@
 // You should have received a copy of the GNU General Public License
 // along with Substrate.  If not, see <http://www.gnu.org/licenses/>.
 
-//! System RPC module errors.
+//! State RPC errors.
 
-use crate::rpc;
-use crate::system::helpers::Health;
+use rpc;
+use crate::errors;
 
-/// System RPC Result type.
+/// State RPC Result type.
 pub type Result<T> = std::result::Result<T, Error>;
 
-/// System RPC errors.
+/// State RPC errors.
 #[derive(Debug, derive_more::Display, derive_more::From)]
 pub enum Error {
+	/// Client error.
+	#[display(fmt="Client error: {}", _0)]
+	Client(Box<dyn std::error::Error + Send>),
 	/// Provided block range couldn't be resolved to a list of blocks.
-	#[display(fmt = "Node is not fully functional: {}", _0)]
-	NotHealthy(Health),
+	#[display(fmt = "Cannot resolve a block range ['{:?}' ... '{:?}]. {}", from, to, details)]
+	InvalidBlockRange {
+		/// Beginning of the block range.
+		from: String,
+		/// End of the block range.
+		to: String,
+		/// Details of the error message.
+		details: String,
+	},
 }
 
-impl std::error::Error for Error {}
+impl std::error::Error for Error {
+	fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+		match self {
+			Error::Client(ref err) => Some(&**err),
+			_ => None,
+		}
+	}
+}
 
-/// Base code for all system errors.
-const BASE_ERROR: i64 = 2000;
+/// Base code for all state errors.
+const BASE_ERROR: i64 = 4000;
 
 impl From<Error> for rpc::Error {
 	fn from(e: Error) -> Self {
 		match e {
-			Error::NotHealthy(ref h) => rpc::Error {
+			Error::InvalidBlockRange { .. } => rpc::Error {
 				code: rpc::ErrorCode::ServerError(BASE_ERROR + 1),
 				message: format!("{}", e),
-				data: serde_json::to_value(h).ok(),
+				data: None,
 			},
+			e => errors::internal(e),
 		}
 	}
 }
