+++ core/rpc/api/src/subscriptions.rs
@@ -20,8 +20,8 @@ use std::sync::{Arc, atomic::{self, AtomicUsize}};
 use log::{error, warn};
 use jsonrpc_pubsub::{SubscriptionId, typed::{Sink, Subscriber}};
 use parking_lot::Mutex;
-use crate::rpc::futures::sync::oneshot;
-use crate::rpc::futures::{Future, future};
+use rpc::futures::sync::oneshot;
+use rpc::futures::{Future, future};
 
 type Id = u64;
