+++ core/rpc/api/src/system/error.rs
@@ -14,53 +14,35 @@
 // You should have received a copy of the GNU General Public License
 // along with Substrate.  If not, see <http://www.gnu.org/licenses/>.
 
-//! State RPC errors.
+//! System RPC module errors.
 
-use client;
-use crate::rpc;
-use crate::errors;
+use rpc;
+use crate::system::helpers::Health;
 
-/// State RPC Result type.
+/// System RPC Result type.
 pub type Result<T> = std::result::Result<T, Error>;
 
-/// State RPC errors.
+/// System RPC errors.
 #[derive(Debug, derive_more::Display, derive_more::From)]
 pub enum Error {
-	/// Client error.
-	Client(client::error::Error),
 	/// Provided block range couldn't be resolved to a list of blocks.
-	#[display(fmt = "Cannot resolve a block range ['{:?}' ... '{:?}]. {}", from, to, details)]
-	InvalidBlockRange {
-		/// Beginning of the block range.
-		from: String,
-		/// End of the block range.
-		to: String,
-		/// Details of the error message.
-		details: String,
-	},
+	#[display(fmt = "Node is not fully functional: {}", _0)]
+	NotHealthy(Health),
 }
 
-impl std::error::Error for Error {
-	fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-		match self {
-			Error::Client(ref err) => Some(err),
-			_ => None,
-		}
-	}
-}
+impl std::error::Error for Error {}
 
-/// Base code for all state errors.
-const BASE_ERROR: i64 = 4000;
+/// Base code for all system errors.
+const BASE_ERROR: i64 = 2000;
 
 impl From<Error> for rpc::Error {
 	fn from(e: Error) -> Self {
 		match e {
-			Error::InvalidBlockRange { .. } => rpc::Error {
+			Error::NotHealthy(ref h) => rpc::Error {
 				code: rpc::ErrorCode::ServerError(BASE_ERROR + 1),
 				message: format!("{}", e),
-				data: None,
+				data: serde_json::to_value(h).ok(),
 			},
-			e => errors::internal(e),
 		}
 	}
 }
